# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Portal Municipal Inteligente - API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenant/config:
    get:
      operationId: getTenantConfig
      tags: [tenant]
      summary: Get tenant configuration
      parameters:
        - name: tenant
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tenant configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantConfig"
...
  /servicos:
    get:
      operationId: listServicos
      tags: [servicos]
      summary: Listar serviços ao cidadão
      parameters:
        - name: categoria
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Lista de serviços
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicoListResponse"
  /servicos/{slug}:
    get:
      operationId: getServico
      tags: [servicos]
      summary: Obter serviço por slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Serviço encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Servico"
        "404":
          description: Serviço não encontrado
...
  /transparencia/orcamento:
    get:
      operationId: getOrcamento
      tags: [transparencia]
      summary: Dados do orçamento municipal
      parameters:
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dados do orçamento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrcamentoResponse"
  /transparencia/despesas:
    get:
      operationId: listDespesas
      tags: [transparencia]
      summary: Listar despesas
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: ano
          in: query
          schema:
            type: integer
        - name: mes
          in: query
          schema:
            type: integer
        - name: secretaria
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de despesas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DespesaListResponse"
  /transparencia/receitas:
    get:
      operationId: listReceitas
      tags: [transparencia]
      summary: Listar receitas
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: ano
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de receitas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceitaListResponse"
  /transparencia/servidores:
    get:
      operationId: listServidores
      tags: [transparencia]
      summary: Listar servidores públicos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: secretaria
          in: query
          schema:
            type: string
        - name: nome
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de servidores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServidorListResponse"
...
components:
  schemas:
...
    TenantConfig:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        slug:
          type: string
        brasao:
          type: string
          nullable: true
        corPrimaria:
          type: string
        corSecundaria:
          type: string
        corTerciaria:
          type: string
        estado:
          type: string
        populacao:
          type: integer
        area:
          type: number
        fundacao:
          type: string
        lema:
          type: string
          nullable: true
        siteUrl:
          type: string
        modulosAtivos:
          type: array
          items:
            type: string
      required:
        - id
        - nome
        - slug
        - corPrimaria
        - corSecundaria
        - corTerciaria
        - estado
        - populacao
        - area
        - fundacao
        - siteUrl
        - modulosAtivos
...
    Servico:
      type: object
      properties:
        id:
          type: string
        titulo:
          type: string
        slug:
          type: string
        descricao:
          type: string
        categoria:
          type: string
        orgao:
          type: string
          nullable: true
        linkExterno:
          type: string
          nullable: true
        requisitos:
          type: array
          items:
            type: string
        documentos:
          type: array
          items:
            type: string
        prazoAtendimento:
          type: string
          nullable: true
        gratuito:
          type: boolean
        online:
          type: boolean
        icone:
          type: string
          nullable: true
      required:
        - id
        - titulo
        - slug
        - descricao
        - categoria
        - requisitos
        - documentos
        - gratuito
        - online
    ServicoListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Servico"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
...
    OrcamentoResponse:
      type: object
      properties:
        ano:
          type: integer
        receitaPrevista:
          type: number
        receitaRealizada:
          type: number
        despesaPrevista:
          type: number
        despesaRealizada:
          type: number
        saldoAtual:
          type: number
        categorias:
          type: array
          items:
            type: object
            properties:
              nome:
                type: string
              valor:
                type: number
              percentual:
                type: number
            required:
              - nome
              - valor
              - percentual
      required:
        - ano
        - receitaPrevista
        - receitaRealizada
        - despesaPrevista
        - despesaRealizada
        - saldoAtual
        - categorias
    Despesa:
      type: object
      properties:
        id:
          type: string
        data:
          type: string
          format: date
        descricao:
          type: string
        secretaria:
          type: string
        categoria:
          type: string
        valor:
          type: number
        beneficiario:
          type: string
          nullable: true
        empenho:
          type: string
          nullable: true
        modalidade:
          type: string
          nullable: true
      required:
        - id
        - data
        - descricao
        - secretaria
        - categoria
        - valor
    DespesaListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Despesa"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    Receita:
      type: object
      properties:
        id:
          type: string
        data:
          type: string
          format: date
        descricao:
          type: string
        fonte:
          type: string
        categoria:
          type: string
        valor:
          type: number
      required:
        - id
        - data
        - descricao
        - fonte
        - categoria
        - valor
    ReceitaListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Receita"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    Servidor:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        cargo:
          type: string
        secretaria:
          type: string
        vinculo:
          type: string
        remuneracao:
          type: number
          nullable: true
        admissao:
          type: string
          format: date
          nullable: true
      required:
        - id
        - nome
        - cargo
        - secretaria
        - vinculo
    ServidorListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Servidor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
...