# lib/api-zod/src/generated/types/despesa.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portal Municipal Inteligente - API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Despesa {
  id: string;
  data: Date;
  descricao: string;
  secretaria: string;
  categoria: string;
  valor: number;
  beneficiario?: string | null;
  empenho?: string | null;
  modalidade?: string | null;
}