# lib/api-zod/src/generated/types/receita.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portal Municipal Inteligente - API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Receita {
  id: string;
  data: Date;
  descricao: string;
  fonte: string;
  categoria: string;
  valor: number;
}